rename swarmauri_matplotlib_tool to swarmauri_tool_matplotlib and update references
