+++ pkgs/standards/swarmauri_tool_matplotlib/pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "swarmauri_matplotlib_tool"
+name = "swarmauri_tool_matplotlib"
 version = "0.6.0.dev1"
 description = "This repository includes an example of a First Class Swarmauri Example."
 authors = ["Jacob Stewart <jacob@swarmauri.com>"]
@@ -50,6 +50,6 @@ log_cli_date_format = "%Y-%m-%d %H:%M:%S"
 asyncio_default_fixture_loop_scope = "function"
 
 [tool.poetry.plugins."swarmauri.tools"]
-MatplotlibTool = "swarmauri_matplotlib_tool:MatplotlibTool"
-MatplotlibCsvTool = "swarmauri_matplotlib_tool:MatplotlibCsvTool"
+MatplotlibTool = "swarmauri_tool_matplotlib:MatplotlibTool"
+MatplotlibCsvTool = "swarmauri_tool_matplotlib:MatplotlibCsvTool"
