+++ pkgs/standards/swarmauri_tool_matplotlib/swarmauri_tool_matplotlib/MatplotlibTool.py
@@ -4,7 +4,6 @@ from typing import List, Literal
 from pydantic import Field
 from swarmauri_base.tools.ToolBase import ToolBase
 from swarmauri_standard.tools.Parameter import Parameter
-from swarmauri_base.tools.ToolBase import ToolBase
 from swarmauri_core.ComponentBase import ComponentBase
