+++ pkgs/standards/swarmauri_tool_matplotlib/tests/unit/MatplotlibCsvTool_unit_test.py
@@ -2,7 +2,7 @@ import os
 from tempfile import NamedTemporaryFile
 
 import pytest
-from swarmauri_matplotlib_tool.MatplotlibCsvTool import MatplotlibCsvTool as Tool
+from swarmauri_tool_matplotlib.MatplotlibCsvTool import MatplotlibCsvTool as Tool
 
 
 @pytest.mark.unit
